+++ README.md
@@ -4,7 +4,7 @@ Pull weather data from Weather Underground and save it to Influxdb
 
 ## Setup
 
-Make a copy of the `config.ini.example` file called `config.ini` and update the values as necessary. Additional Location sections can be added to pull as many locations as you'd like. By default the cron job runs the scrape every 5 minutes. This can be adjusted in the `crontab` file.
+Make a copy of the `config.ini.example` file called `config.ini` and update the values as necessary. Additional Location sections can be added to pull as many locations as you'd like. By default the job runs the scrape every 5 minutes. This can be adjusted using the `interval` parameter in `config.ini`.
 
 ## Run
