+++ docker-compose.yaml
@@ -12,4 +12,3 @@ services:
       # - logs/cron.log:/var/log/cron.log  # Bind-mount for cron logs (optional)
     environment:
       POETRY_VIRTUALENVS_CREATE: "false"  # Set Poetry environment variables if necessary
-    command: cron -f  # Start cron as the entry point
