+++ pyproject.toml
@@ -13,6 +13,7 @@ WunderWeather = "^1.1.0"
 easydict = "^1.13"
 arrow = "^1.3.0"
 influxdb-client = "^1.47.0"
+schedule = "^1.2.2"
 
 
 [tool.poetry.group.dev.dependencies]
